split spark 3.1.1 into a base shim to easily copy to other 3.1.x
